requirements/test.txt
mock
coverage
django-nose
coveralls
django-webtest
tox

# Open edX
analytics-python==1.2.9
babel==1.3
bleach>=3.1.1
boto==2.39.0
celery==3.1.25
django-config-models==0.2.2
django-countries==4.6.1
django-crum==0.7.3
django-model-utils==3.0.0
django-oauth-toolkit==1.1.3
django-ratelimit-backend==1.1.1
django-require==1.0.11
django-simple-history==2.7.0
django-storages==1.4.1
django-waffle==0.12.0
django==1.11.21
dogapi==1.2.1
edx-ace==0.1.10
edx-ccx-keys==0.2.1
edx-celeryutils==0.2.7
edx-completion==2.0.0
edx-django-oauth2-provider==1.3.5
edx-milestones==0.1.13
edx-oauth2-provider==1.2.2
edx-opaque-keys[django]==0.4.4
edx-proctoring==1.5.7
edx-rest-api-client==1.9.2
edx-search==1.2.1
edx-submissions==2.0.12
edx-user-state-client==1.0.4
event-tracking==0.2.7
fs==2.0.18
git+https://github.com/edx/MongoDBProxy.git@25b99097615bda06bd7cdfe5669ed80dc2a7fed0#egg=MongoDBProxy==0.1.0
git+https://github.com/edx/codejail.git@a320d43ce6b9c93b17636b2491f724d9e433be47#egg=codejail==0.0
git+https://github.com/edx/django-celery.git@756cb57aad765cb2b0d37372c1855b8f5f37e6b0#egg=django-celery==3.2.1+edx.2
git+https://github.com/edx/django-rest-framework-oauth.git@0a43e8525f1e3048efe4bc70c03de308a277197c#egg=djangorestframework-oauth==1.1.1
git+https://github.com/jazzband/django-pipeline.git@d068a019169c9de5ee20ece041a6dea236422852#egg=django-pipeline==1.5.3
html5lib==1.0.1
jsonfield==2.0.2
lazy==1.1
mako==1.0.2
nltk==3.4
numpy==1.6.2
oauthlib==2.1.0
path.py==8.2.1
pillow>=6.2.0
pycontracts==1.7.1
pycountry==1.20
pycrypto>=2.6
pycryptodomex==3.4.7
pyjwkest==1.3.2
pymongo==2.9.1
sailthru-client==2.2.3
scipy==0.14.0
shapely==1.2.16
sortedcontainers==0.9.2
sqlparse==0.2.4
user-util==0.1.5
xblock==1.2.2
